Rules with equal literals are now allowed
This is a fix for bug #1829204: Repeated literal in query fails with magic sets

The sip now labels the edges with exactly with the variables, which are
passed from the source to the target.

All unconnected literals are now added to the sip, too.

Deleted much unnecessary code and cleaned up some other.

Because of the new sip more labeled rules are created, but no rules for magic
predicates with only free adornments.